third_party/blink/web_tests/fast/scroll-snap/snaps-after-keyboard-scrolling-1.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1" />
<link rel="stylesheet" href="resources/simple-snap.css">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/gesture-util.js"></script>

<div id='scroller'>
  <div id="space"></div>
  <div class="snap left top" id="top-left"></div>
  <div class="snap right top" id="top-right"></div>
  <div class="snap left bottom" id="bottom-left"></div>
</div>

<script>
var scroller = document.getElementById("scroller");
var topLeft = document.getElementById("top-left");
var topRight = document.getElementById("top-right");

function keyPressAndScroll(key) {
  return new Promise((resolve, reject) => {
    if (window.eventSender) {
      const scrollend = waitForScrollendEvent(scroller);
      eventSender.keyDown(key);
      scrollend.then(resolve);
    }
    else {
      reject('This test requires window.eventSender');
    }
  });
}

window.onload = async () => {
  await waitForCompositorCommit();
  await mouseClickOn(10, 10);

  promise_test (async () => {
    await waitForScrollReset(scroller);
    await keyPressAndScroll("ArrowDown");
    assert_approx_equals(scroller.scrollTop, 400, 1);
  }, "Snaps to bottom-left after pressing ArrowDown");

  promise_test (async () => {
    await waitForScrollReset(scroller, /*x=*/0, /*y=*/400);
    await keyPressAndScroll("ArrowUp");
    assert_approx_equals(scroller.scrollTop, 0, 1);
  }, "Snaps to top-left after pressing ArrowUp");

  promise_test (async () => {
    await waitForScrollReset(scroller);
    await keyPressAndScroll("ArrowRight");
    assert_approx_equals(scroller.scrollLeft, 400, 1);
  }, "Snaps to top-right after pressing ArrowRight");

  promise_test (async () => {
    await waitForScrollReset(scroller, /*x=*/400, /*y=*/0);
    await keyPressAndScroll("ArrowLeft");
    assert_approx_equals(scroller.scrollLeft, 0, 1);
  }, "Snaps to top-left after pressing ArrowLeft");
};


</script>
